move finetune common scripts to refact_utils
