+++ refact_utils/finetune/utils.py
@@ -6,8 +6,8 @@ import time
 from pathlib import Path
 
 from refact_utils.scripts import env
-from self_hosting_machinery.finetune.configuration.finetune_filtering_defaults import finetune_filtering_defaults
-from self_hosting_machinery.finetune.configuration.finetune_train_defaults import finetune_train_defaults
+from refact_utils.finetune.filtering_defaults import finetune_filtering_defaults
+from refact_utils.finetune.train_defaults import finetune_train_defaults
 
 from typing import Any, Dict, Optional, Callable, Union
