+++ refact_utils/scripts/best_lora.py
@@ -3,8 +3,8 @@ import os
 import json
 
 from refact_utils.scripts import env
-from self_hosting_machinery.finetune.utils.finetune_utils import get_run_model_name  # REFACTORME
-from self_hosting_machinery.finetune.utils.finetune_utils import default_finetune_model  # REFACTORME
+from refact_utils.finetune.utils import get_run_model_name
+from refact_utils.finetune.utils import default_finetune_model
 
 from typing import Dict, Optional
