+++ refact_webgui/webgui/selfhost_model_assigner.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass, field
 
 from refact_utils.scripts import env
 from refact_utils.scripts.best_lora import find_best_lora
-from self_hosting_machinery.finetune.utils.finetune_utils import get_active_loras  # REFACTORME
+from refact_utils.finetune.utils import get_active_loras
 from refact_webgui.webgui.selfhost_webutils import log
 from known_models_db.refact_known_models import models_mini_db
 from known_models_db.refact_toolbox_db import modelcap_records
