+++ refact_webgui/webgui/tab_upload.py
@@ -12,9 +12,8 @@ from pydantic import BaseModel, Required
 from fastapi import APIRouter, Request, Query, UploadFile, HTTPException
 from fastapi.responses import Response, JSONResponse, StreamingResponse
 
-from self_hosting_machinery.finetune.utils.finetune_utils import get_prog_and_status_for_ui  # REFACTORME
-
 from refact_utils.scripts import env
+from refact_utils.finetune.utils import get_prog_and_status_for_ui
 from refact_webgui.webgui.selfhost_webutils import log
