+++ self_hosting_machinery/inference/lora_loader_mixin.py
@@ -10,9 +10,9 @@ from safetensors.torch import load_file
 
 from refact_utils.scripts import env
 from refact_utils.scripts import best_lora
+from refact_utils.finetune.utils import get_active_loras
 from self_hosting_machinery.finetune.modelling.lora import LoraMixin
 from self_hosting_machinery.finetune.modelling.utils import map_model_specific_params
-from self_hosting_machinery.finetune.utils.finetune_utils import get_active_loras
 
 log = logging.getLogger("MODEL").info
